moving LoadingIndicator functionality into its own component
